more tests

+++ VirtualVpn/Log.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
+// ReSharper disable UnusedMember.Global
 
 namespace VirtualVpn;
 
@@ -39,13 +40,13 @@ public static class Log
 
     public static void Crypto(string msg)
     {
-        if (_level < LogLevel.Crypto) return;
+        if ((int)_level < (int)LogLevel.Crypto) return;
         Console.WriteLine(msg);
     }
 
     public static void Trace(string msg)
     {
-        if (_level < LogLevel.Trace) return;
+        if ((int)_level < (int)LogLevel.Trace) return;
         
         Console.Write("                       "); // same spacing as timestamp
         Console.WriteLine(msg);
@@ -53,7 +54,7 @@ public static class Log
     
     public static void Trace(string msg, Func<string> more)
     {
-        if (_level < LogLevel.Trace) return;
+        if ((int)_level < (int)LogLevel.Trace) return;
         
         Console.Write("                       "); // same spacing as timestamp
         Console.Write(msg);
@@ -62,7 +63,7 @@ public static class Log
 
     public static void Debug(string msg, Func<IEnumerable<string>>? subLines = null)
     {
-        if (_level < LogLevel.Debug) return;
+        if ((int)_level < (int)LogLevel.Debug) return;
         Timestamp();
         
         Console.WriteLine(msg);
@@ -78,7 +79,7 @@ public static class Log
     
     public static void Debug(IEnumerable<string> messages)
     {
-        if (_level < LogLevel.Debug) return;
+        if ((int)_level < (int)LogLevel.Debug) return;
         Timestamp();
         
         foreach (var msg in messages)
@@ -92,7 +93,7 @@ public static class Log
 
     public static void Info(string msg)
     {
-        if (_level < LogLevel.Info) return;
+        if ((int)_level < (int)LogLevel.Info) return;
         Timestamp();
         
         Console.WriteLine(msg);
@@ -100,7 +101,7 @@ public static class Log
     
     public static void Warn(string msg)
     {
-        if (_level < LogLevel.Warning) return;
+        if ((int)_level < (int)LogLevel.Warning) return;
         Timestamp();
         
         Console.WriteLine(msg);
@@ -108,7 +109,7 @@ public static class Log
     
     public static void Error(string msg)
     {
-        if (_level < LogLevel.Error) return;
+        if ((int)_level < (int)LogLevel.Error) return;
         Timestamp();
         
         Console.WriteLine(msg);
@@ -117,10 +118,10 @@ public static class Log
 
     public static void Error(string message, Exception ex)
     {
-        if (_level < LogLevel.Error) return;
+        if ((int)_level < (int)LogLevel.Error) return;
         Timestamp();
 
-        if (_level >= LogLevel.Debug)
+        if ((int)_level >= (int)LogLevel.Debug)
         {
             Console.WriteLine(message + ": " + ex); // full trace with debug
         }
